web/requirements.txt
fastapi==0.89.1
better-ipc==2.0.1
python-dotenv==0.21.0
uvicorn==0.20.0
discord.py==2.1.0
Jinja2==3.1.2